pyproject.toml
[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[project]
name = "reddress"
version = "0.3.0"
authors = [
    { name="Jayaharisai Tothala", email="jayaharisai1212@gmail.com" }
]
description = "Reddress"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "openai>=1.0.0",
    "click>=8.0.0",
    "toml>=0.10.0",
    "questionary>=2.0.0",
    "prompt-toolkit>=3.0.0",
    "rich>=13.0.0",
    "fastapi",
    "uvicorn",
    "jinja2",
    "docker",
    "chromadb>=0.4.0",              # ChromaDB vector store
    "qdrant-client>=1.7.0",         # Qdrant vector store
    "pymilvus>=2.3.0",              # Milvus vector store
    "PyPDF2>=3.0.0",                # PDF parsing
    "python-docx>=1.0.0",           # Word document parsing
    "sentence-transformers>=2.2.0",  # Optional: local embeddings
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1.0",
]


[project.scripts]
reddress = "reddress.cli:main"
